+++ vpin-studio-rest-client/src/main/java/de/mephisto/vpin/restclient/GameType.java
@@ -1,0 +1,5 @@
+package de.mephisto.vpin.restclient;
+
+public enum GameType {
+  SS, EM, Original
+}