+++ RELEASE_NOTES.md
@@ -9,3 +9,8 @@
 **Tables Management: PinUP Popper Media**
 
 - Replace "on-hover" upload button with the new asset manager button. The button initializes the asset manager with the selected table and screen.
+
+**Table Management: PinUP Popper Table Data**
+
+- Added "Game Type", "Game Version (File Version)" and "Game Theme" fields.
+-